pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradedesk"
version = "0.1.0"
description = "Trading infrastructure library for IG Markets"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Darren Davison", email = "darren@davisononline.org"}
]
dependencies = [
    "aiohttp>=3.9.0",
    "requests>=2.31",
    "lightstreamer-client-lib>=2.0",
    "python-dotenv>=1.0",
    "numpy>=2.4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["tradedesk"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tradedesk*"]
exclude = ["tests*", "examples*"]
